Modularize tracers
