+++ openfisca_core/tracers/__init__.py
@@ -1,0 +1,30 @@
+# Transitional imports to ensure non-breaking changes.
+# Could be deprecated in the next major release.
+#
+# How imports are being used today:
+#
+#   >>> from openfisca_core.module import symbol
+#
+# The previous example provokes cyclic dependency problems
+# that prevent us from modularizing the different components
+# of the library so to make them easier to test and to maintain.
+#
+# How could them be used after the next major release:
+#
+#   >>> from openfisca_core import module
+#   >>> module.symbol()
+#
+# And for classes:
+#
+#   >>> from openfisca_core import module
+#   >>> module.Symbol()
+#
+# See: https://www.python.org/dev/peps/pep-0008/#imports
+
+from .tracing_parameter_node_at_instant import TracingParameterNodeAtInstant  # noqa: F401
+from .trace_node import TraceNode  # noqa: F401
+from .flat_trace import FlatTrace  # noqa: F401
+from .full_tracer import FullTracer  # noqa: F401
+from .simple_tracer import SimpleTracer  # noqa: F401
+from .computation_log import ComputationLog  # noqa: F401
+from .performance_log import PerformanceLog  # noqa: F401
